feat: :sparkles: Finished working on Yt component
