+++ apps/web/app/yt/page.tsx
@@ -1,15 +1,22 @@
 import React from 'react';
 
-import YoutubePlaylist from './Youtube';
-import getPlaylistSnippet from './FetchPlaylist';
+
+// import YoutubePlaylist from './Youtube';
+// import getPlaylistSnippet from './FetchPlaylist';
+
+// import  getPlaylistSnippet from "@youtube/core/FetchPlaylist"
+// import YoutubePlaylist from '@youtube/ui/Youtube';
+
+import getPlaylistSnippet from '@youtube/src/core/FetchPlaylist';
+import YoutubePlaylist from '@youtube/src/ui/Youtube';
 
 
 const Home = async () => {
   const playlistId = 'PLee3gqXJQrFXj3VqMruL_uC-7pb21cG3g';
   const API_Key = process.env.GOOGLE_API_KEY;
-  
-  const PlayListData =  await getPlaylistSnippet(playlistId,API_Key)
 
+
+  const PlayListData =  await getPlaylistSnippet(playlistId, API_Key)
 
   return (
     <YoutubePlaylist
@@ -20,6 +27,7 @@ const Home = async () => {
       width={"550px"}
       title={"Top GSOC 2022 Videos"}
     />
+
   );
 };
