+++ apps/web/package.json
@@ -31,7 +31,7 @@
     "react-dom": "^18.2.0",
     "tailwind-merge": "^1.9.1",
     "tailwindcss-animate": "^1.0.5",
-    "youtube": "workspace:*"
+    "@youtube": "workspace:*"
   },
   "devDependencies": {
     "@types/node": "^18.11.17",
