+++ packages/youtube/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "youtube",
+  "name": "@youtube",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
